backend/requirements.txt
# MongoDB Version
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.1
motor==3.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.1
numpy==1.24.4
opencv-python-headless==4.9.0.80
pillow==10.2.0
torch==2.1.0
torchvision==0.16.0
trimesh==4.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2